Normalize integer fields for OpenAPI compliance (wind_speed, gust_speed, ceiling, etc.)
- Source: SWOB adapter casts wind/ceiling to int; calculator returns int for pressure_altitude
- API: toApiInteger() in format layer enforces spec for Public, Internal, and History endpoints
- Fixes CYAV and other SWOB airports returning floats (e.g. 11.66...) instead of integers

+++ lib/weather/calculator.php
@@ -135,8 +135,8 @@ function calculatePressureAltitude($weather, $airport) {
     
     // PA = Station Elevation + (29.92 - Altimeter) × 1000
     $pressureAlt = $stationElevation + (29.92 - $pressureInHg) * 1000;
-    
-    return round($pressureAlt);
+
+    return (int) round($pressureAlt);
 }
 
 /**
